refactor(notebook): extract export reader
